refactor: switch addressValidate to account utils
