+++ src/address/types.ts
@@ -9,5 +9,3 @@ export type DecodedAddress = {
     viewPublicKey: Buffer;
     spendPublicKey: Buffer;
 }
-
-export type AddressValidateResult = string | true;
